Baks Development

+++ Messenger/WbCardNew/CardCreateHandler.php
@@ -58,7 +58,7 @@ use BaksDev\Wildberries\Api\Token\Card\WildberriesCards\Card;
 use BaksDev\Wildberries\Api\Token\Prices\PricesInfo\PricesInfo;
 use BaksDev\Wildberries\Api\Token\Stocks\GetStocks\Stocks;
 use BaksDev\Wildberries\Api\Token\Stocks\GetStocks\WildberriesStocks;
-use BaksDev\Wildberries\Api\Token\Warehouse\PartnerWildberries\PartnerWarehouses;
+use BaksDev\Wildberries\Api\Token\Warehouse\ProfileWarehouses\ProfileWarehousesClient;
 use BaksDev\Wildberries\Products\Entity\Cards\WbProductCard;
 use BaksDev\Wildberries\Products\Entity\Settings\Event\WbProductSettingsEvent;
 use BaksDev\Wildberries\Products\Repository\Settings\ProductSettingsByParentAndName\ProductSettingsByParentAndNameInterface;
@@ -85,25 +85,27 @@ final class CardCreateHandler
     private ProductSettingsByParentAndNameInterface $productSettingsByParentAndName;
     private OfferByCategoryInterface $offerByCategory;
     private VariationByOfferInterface $variationByOffer;
-    private LoggerInterface $logger;
+    private LoggerInterface $messageDispatchLogger;
     private WildberriesCardImage $wildberriesCardImage;
     private PricesInfo $wildberriesPricesInfo;
     private WildberriesStocks $wildberriesStocks;
-    private PartnerWarehouses $wildberriesPartnerWarehouses;
+
     private ProductHandler $productHandler;
     private WbProductCardHandler $wbProductCardHandler;
     private EntityManagerInterface $entityManager;
+
+    private ProfileWarehousesClient $profileWarehousesClient;
 
     public function __construct(
         #[TaggedIterator('baks.reference.choice')] iterable $reference,
         ProductSettingsByParentAndNameInterface $productSettingsByParentAndName,
         OfferByCategoryInterface $offerByCategory,
         VariationByOfferInterface $variationByOffer,
-        LoggerInterface $wildberriesProductsLogger,
+        LoggerInterface $messageDispatchLogger,
         WildberriesCardImage $wildberriesCardImage,
         PricesInfo $wildberriesPricesInfo,
         WildberriesStocks $wildberriesStocks,
-        PartnerWarehouses $wildberriesPartnerWarehouses,
+        ProfileWarehousesClient $profileWarehousesClient,
         ProductHandler $productHandler,
         WbProductCardHandler $wbProductCardHandler,
         EntityManagerInterface $entityManager
@@ -113,19 +115,21 @@ final class CardCreateHandler
         $this->offerByCategory = $offerByCategory;
         $this->variationByOffer = $variationByOffer;
         $this->reference = $reference;
-        $this->logger = $wildberriesProductsLogger;
+        $this->messageDispatchLogger = $messageDispatchLogger;
         $this->wildberriesCardImage = $wildberriesCardImage;
         $this->wildberriesPricesInfo = $wildberriesPricesInfo;
         $this->wildberriesStocks = $wildberriesStocks;
-        $this->wildberriesPartnerWarehouses = $wildberriesPartnerWarehouses;
+
         $this->productHandler = $productHandler;
         $this->wbProductCardHandler = $wbProductCardHandler;
         $this->entityManager = $entityManager;
+        $this->profileWarehousesClient = $profileWarehousesClient;
     }
 
     public function __invoke(Card $Card): void
     {
-        $this->Card = $Card->getCardDetail();
+
+        $this->Card = $Card;
 
         /* Создаем системный продукт */
         $this->createProduct();
@@ -137,11 +141,16 @@ final class CardCreateHandler
         $this->createProductRootCategory($WbProductsSettingsDTO->getSettings());
 
 
-        /** Характеристики продукции */
-        /** @var WbProductSettingsPropertyDTO $property */
+        /**
+         * Характеристики продукции
+         * @var WbProductSettingsPropertyDTO $property
+         */
+
         foreach($WbProductsSettingsDTO->getProperty() as $property)
         {
-            $characteristics = $Card->getCharacteristic($property->getType());
+
+            $propertyType = mb_strtolower($property->getType());
+            $characteristics = $Card->getCharacteristic($propertyType);
 
             if($characteristics)
             {
@@ -149,106 +158,130 @@ final class CardCreateHandler
             }
         }
 
-        if($Card->getColor())
+
+        /**
+         * Торговое предложение продукции
+         */
+        $ProductOfferValue = null;
+
+        $SettingsCategoryOffers = $this->getSettingsCategoryOffers($WbProductsSettingsDTO->getSettings());
+
+        if($SettingsCategoryOffers)
         {
-            /**
-             * Торговое предложение продукции
-             */
-
-            $SettingsCategoryOffers = $this->getSettingsCategoryOffers($WbProductsSettingsDTO->getSettings());
-            $ProductOffer = $this->createProductOffer($SettingsCategoryOffers?->getId());
-            $ProductOffer->setValue($Card->getColor());
-
-            /* Если свойство из справочника */
-            if($SettingsCategoryOffers && $SettingsCategoryOffers->getReference())
-            {
-                $this->setReference($ProductOffer, $SettingsCategoryOffers->getReference(), $Card->getColor());
-            }
-
-            /* Если торговое предложение с артикулом */
-            if($SettingsCategoryOffers->getArticle())
-            {
-                $ProductOffer->setArticle($Card->getArticle());
-                $this->ProductDTO->getInfo()->setArticle($Card->getBaseArticle());
-            }
-
-            /* Если торговое предложение с возможностью загрузки пользовательского изображение */
-            if($SettingsCategoryOffers->getImage())
-            {
-                $this->createMediaFile($ProductOffer, ProductOfferImage::TABLE, ProductOfferImageCollectionDTO::class);
-            }
-
-            /* Если торговое предложение с ценой  */
-            if($SettingsCategoryOffers->getPrice())
-            {
-                $this->createProductOfferPrice($ProductOffer);
-            }
-
-            /* Если торговое предложение с количественным учетом */
-            if($SettingsCategoryOffers->getQuantitative())
-            {
-                $this->createProductOfferQuantity($ProductOffer, $Card->getCurrentBarcode());
-            }
-
 
             /**
-             * Множественный варианты продукции
+             * Поиск по названию в характеристиках
              */
-
-            $SettingsCategoryVariation = $this->getSettingsCategoryVariation($WbProductsSettingsDTO->getSettings());
-
-
-            if(!$SettingsCategoryVariation || !$Card->isOffers())
+            foreach($SettingsCategoryOffers->getTranslate() as $trans)
             {
-                /** Всегда создаем по умолчанию множественный вариант с нулевым значением */
-                $ProductVariation = $this->createProductVariation($ProductOffer, $Card->getCurrentBarcode(), $SettingsCategoryVariation?->getId());
-                $ProductVariation->setValue($Card->getCurrentValue());
-            }
-            else
-            {
-                foreach($this->Card->getOffersCollection() as $barcode => $value)
+                $searchOffer = mb_strtolower($trans->getName());
+
+                $ProductOfferValue = $Card->getCharacteristic($searchOffer);
+
+                if($ProductOfferValue)
                 {
-                    $ProductVariation = $this->createProductVariation($ProductOffer, $barcode, $SettingsCategoryVariation->getId());
-                    $ProductVariation->setValue($value);
+                    break;
+                }
+            }
+
+            if($ProductOfferValue)
+            {
+                $ProductOffer = $this->createProductOffer($SettingsCategoryOffers?->getId());
+                $ProductOffer->setValue($ProductOfferValue);
+
+                /* Если свойство из справочника */
+                if($SettingsCategoryOffers->getReference())
+                {
+                    $this->setReference($ProductOffer, $SettingsCategoryOffers->getReference(), $ProductOfferValue);
+                }
+
+                /* Если торговое предложение с артикулом */
+                if($SettingsCategoryOffers->getArticle())
+                {
+                    $ProductOffer->setArticle($Card->getArticle());
+                    $this->ProductDTO->getInfo()->setArticle($Card->getBaseArticle());
+                }
+
+                /* Если торговое предложение с возможностью загрузки пользовательского изображение */
+                if($SettingsCategoryOffers->getImage())
+                {
+                    $this->createMediaFile($ProductOffer, ProductOfferImage::TABLE, ProductOfferImageCollectionDTO::class);
+                }
+
+                /* Если торговое предложение с ценой  */
+                if($SettingsCategoryOffers->getPrice())
+                {
+                    $this->createProductOfferPrice($ProductOffer);
+                }
+
+                /* Если торговое предложение с количественным учетом */
+                if($SettingsCategoryOffers->getQuantitative())
+                {
+                    $this->createProductOfferQuantity($ProductOffer, $Card->getCurrentBarcode());
+                }
 
 
-                    /* Если множественное свойство из справочника */
-                    if($SettingsCategoryVariation->getReference())
+                /**
+                 * Множественный варианты продукции
+                 */
+
+                $SettingsCategoryVariation = $this->getSettingsCategoryVariation($WbProductsSettingsDTO->getSettings());
+
+
+                if(!$SettingsCategoryVariation || !$Card->isOffers())
+                {
+                    /** Всегда создаем по умолчанию множественный вариант с нулевым значением */
+                    $ProductVariation = $this->createProductVariation($ProductOffer, $Card->getCurrentBarcode(), $SettingsCategoryVariation?->getId());
+                    $ProductVariation->setValue($Card->getCurrentValue());
+                }
+                else
+                {
+                    foreach($this->Card->getOffersCollection() as $barcode => $value)
                     {
-                        $this->setReference($ProductVariation, $SettingsCategoryVariation->getReference(), $value);
-                    }
+                        $ProductVariation = $this->createProductVariation($ProductOffer, $barcode, $SettingsCategoryVariation->getId());
+                        $ProductVariation->setValue($value);
 
-                    /* Если множественный вариант с артикулом */
-                    if($SettingsCategoryVariation->getArticle())
-                    {
-                        $ProductVariation->setArticle($Card->getArticle());
-                        $this->ProductDTO->getInfo()->setArticle($Card->getBaseArticle());
-                    }
 
-                    /* Если множественный вариант с возможностью загрузки пользовательского изображение */
-                    if($SettingsCategoryVariation->getImage())
-                    {
-                        $this->createMediaFile($ProductVariation, ProductVariationImage::TABLE, ProductVariationImageCollectionDTO::class);
-                    }
+                        /* Если множественное свойство из справочника */
+                        if($SettingsCategoryVariation->getReference())
+                        {
+                            $this->setReference($ProductVariation, $SettingsCategoryVariation->getReference(), $value);
+                        }
 
-                    /* Если множественный вариант с ценой  */
-                    if($SettingsCategoryVariation->getPrice())
-                    {
-                        $this->createProductVariationPrice($ProductVariation);
-                    }
+                        /* Если множественный вариант с артикулом */
+                        if($SettingsCategoryVariation->getArticle())
+                        {
+                            $ProductVariation->setArticle($Card->getArticle());
+                            $this->ProductDTO->getInfo()->setArticle($Card->getBaseArticle());
+                        }
 
-                    /* Если множественный вариант с количественным учетом */
-                    if($SettingsCategoryVariation->getQuantitative())
-                    {
-                        $this->createProductVariationQuantity($ProductVariation, $barcode);
+                        /* Если множественный вариант с возможностью загрузки пользовательского изображение */
+                        if($SettingsCategoryVariation->getImage())
+                        {
+                            $this->createMediaFile($ProductVariation, ProductVariationImage::TABLE, ProductVariationImageCollectionDTO::class);
+                        }
+
+                        /* Если множественный вариант с ценой  */
+                        if($SettingsCategoryVariation->getPrice())
+                        {
+                            $this->createProductVariationPrice($ProductVariation);
+                        }
+
+                        /* Если множественный вариант с количественным учетом */
+                        if($SettingsCategoryVariation->getQuantitative())
+                        {
+                            $this->createProductVariationQuantity($ProductVariation, $barcode);
+                        }
                     }
                 }
+
             }
 
         }
 
 
         $Product = $this->productHandler->handle($this->ProductDTO);
+
 
         if($Product instanceof Product)
         {
@@ -270,13 +303,16 @@ final class CardCreateHandler
                 }
 
                 $DeleteProduct = $this->entityManager->getRepository(Product::class)->find($Product->getId());
-                $this->entityManager->remove($DeleteProduct);
+
+                if($DeleteProduct)
+                {
+                    $this->entityManager->remove($DeleteProduct);
+                }
 
                 $this->entityManager->flush();
             }
-
-
         }
+
     }
 
     public function createProduct(): void
@@ -287,7 +323,8 @@ final class CardCreateHandler
         $ProductInfo = $this->ProductDTO->getInfo();
         $ProductInfo->setUrl(uniqid('', false));
         $ProductInfo->setProfile($this->Card->getProfile());
-        $ProductInfo->setArticle($this->Card->getColor() ? $this->Card->getBaseArticle() : $this->Card->getArticle());
+        $ProductInfo->setArticle($this->Card->getOffersCollection()->count() > 1 ? $this->Card->getBaseArticle() : $this->Card->getArticle());
+
         $this->ProductDTO->setInfo($ProductInfo);
 
 
@@ -303,12 +340,17 @@ final class CardCreateHandler
         foreach($this->ProductDTO->getDescription() as $ProductDescriptionDTO)
         {
             $ProductDescriptionDTO->setPreview($this->Card->getName());
-            $ProductDescriptionDTO->setPreview($this->Card->getDescription());
+            $ProductDescriptionDTO->setDescription($this->Card->getDescription());
         }
 
+
         /** Стоимость продукции Wildberries всегда общая для всех торговых предложений */
-        $wildberriesPrices = $this->wildberriesPricesInfo->profile($this->Card->getProfile())->prices();
+        $wildberriesPrices = $this->wildberriesPricesInfo
+            ->profile($this->Card->getProfile())
+            ->prices($this->Card->getNomenclature());
+
         $Price = $wildberriesPrices->getPriceByNomenclature($this->Card->getNomenclature());
+
 
         $PriceDTO = new PriceDTO();
         $PriceDTO->setPrice($Price?->getPrice());
@@ -426,7 +468,7 @@ final class CardCreateHandler
                 {
                     $error = sprintf('%s: В библиотеке отсутствует значение', $this->Card->getProfile());
 
-                    $this->logger->warning($error,
+                    $this->messageDispatchLogger->warning($error,
                         [
                             'class' => $referenceClass,
                             'value' => $this->Card->getColor(),
@@ -485,6 +527,7 @@ final class CardCreateHandler
         /** @var WbProductSettingsEvent $WbProductSettingsEvent */
         $WbProductSettingsEvent = $this->productSettingsByParentAndName->get($category);
         return $WbProductSettingsEvent?->getDto(WbProductsSettingsDTO::class);
+
     }
 
     private function createProductOfferPrice(ProductOffersCollectionDTO $ProductOffer): void
@@ -514,7 +557,7 @@ final class CardCreateHandler
     /** Получаем все остатки указанной продукции */
     public function getWildberriesStocks(): ?Stocks
     {
-        $partnerWarehouses = $this->wildberriesPartnerWarehouses
+        $partnerWarehouses = $this->profileWarehousesClient
             ->profile($this->Card->getProfile())
             ->warehouses();
 
@@ -556,7 +599,10 @@ final class CardCreateHandler
     }
 
 
-    private function createProductVariationQuantity(ProductOffersVariationCollectionDTO $ProductVariation, int|string $barcode): void
+    private function createProductVariationQuantity(
+        ProductOffersVariationCollectionDTO $ProductVariation,
+        int|string $barcode
+    ): void
     {
         /** Остатки продукции по всем баркодам (Wildberries Api) */
         $wildberriesStocks = $this->getWildberriesStocks();
@@ -573,13 +619,12 @@ final class CardCreateHandler
         }
     }
 
-    private function createProductProperty(ProductCategorySectionFieldUid $field, mixed $value) : void
+    private function createProductProperty(ProductCategorySectionFieldUid $field, mixed $value): void
     {
         $ProductPropertyDto = new PropertyCollectionDTO();
         $ProductPropertyDto->setField($field);
         $ProductPropertyDto->setValue((string) $value);
         $this->ProductDTO->addProperty($ProductPropertyDto);
     }
-
 
 }
